Extracting classes to separate file
